+++ server/index.js
@@ -11,7 +11,11 @@ import YAML from 'yamljs';
 import swaggerUi from 'swagger-ui-express';
 import swaggerSpec from './config/swagger';
 import routes from './routes/index';
-import { facebookStrategy, googleStrategy, twitterStrategy } from './socialMediaService/passport';
+import {
+  facebookStrategy,
+  googleStrategy,
+  twitterStrategy
+} from './socialMediaService/passport';
 
 dotenv.config();
 const isProduction = process.env.NODE_ENV === 'production';
@@ -34,7 +38,6 @@ app.use(passport.initialize());
 passport.use('facebook', facebookStrategy);
 passport.use('google', googleStrategy);
 passport.use('twitter', twitterStrategy);
-
 
 if (!isProduction) {
   app.use(errorhandler());
